Separates OperationalCarbon calculation from IchnosCF script
